feat: phase-B cutover — the sandbox belt model IS the drive sim
One locomotion source of truth: src/track/forces.rs (pure sandbox force
law — support columns, slip-ellipse traction, governor belt dynamics),
consumed by the new game adapter src/track/sim.rs AND the sandbox
model4 shim. src/driving/ (placeholder raycast sim) deleted outright —
no forks, no conditionals, no hold/bristle transplant (owner call:
nothing structural carries over; slope-parking creep is an accepted
known gap, future hill-hold = per-element bristle of THIS model).

- TrackDrive {throttle, steer, [speed, phase]×2}: owner-predicted,
  replicated, rolled back (LinearVelocity pattern); hblt hash stream;
  WIRE_SURFACE/WIRE_TYPES re-pinned, PROTOCOL_REV 12
- TrackContacts telemetry feeds debug arrows + grounded diagnostics
- Tiger spec: powertrain/support scaled from sandbox tuning (stiff
  governor), drivetrain/suspension blocks die; vehicle proxies get
  Friction::ZERO (min-combine) so Avian adds no phantom grip
- View consumes TrackDrive speed/phase directly (pose-delta belt
  derivation deleted)
- Trace schema v2: per-side belt/loads/gnd; anc/ancm retired;
  divergence + jitter analyzers retargeted
- Legacy residue sweep: stale suspension/anchor/DriveState wording
  purged from docs and comments; headless conflict filter repointed
  at track::sim system names

Gates: 223 tests green, clippy zero, sandbox reversal harness
bit-identical to the commit-B baseline, MP smoke (server+client)
clean — rig bound, no panics, no errors.

+++ src/track/forces.rs
@@ -66,8 +66,10 @@ pub struct ForceParams {
 pub struct SideState {
     /// Belt surface speed (m/s).
     pub speed: f32,
-    /// Belt travel (m) — advects the force stations; also the view's scroll phase.
-    pub phase: f32,
+    /// Belt travel (m) — advects the force stations; also the view's scroll phase. `f64`: it
+    /// grows unbounded and an f32 loses sub-pitch precision within a long match's driving
+    /// distance (codex phase-B finding 8).
+    pub phase: f64,
 }
 
 /// One side's per-tick input.
@@ -162,7 +164,11 @@ pub fn step_side<O: TerrainOracle>(
     let mut belt_reaction = 0.0;
 
     let pitch = polyline_len(input.loop_pts) / input.count.max(1) as f32;
-    let mut stations = resample(input.loop_pts, pitch, state.phase.rem_euclid(pitch));
+    let mut stations = resample(
+        input.loop_pts,
+        pitch,
+        state.phase.rem_euclid(f64::from(pitch)) as f32,
+    );
     stations.truncate(input.count);
     let n = stations.len();
     if n < 3 {
@@ -292,6 +298,6 @@ pub fn step_side<O: TerrainOracle>(
     let engine = (params.governor_gain * (target - belt_speed)).clamp(-avail, avail);
     let next = belt_speed + (engine - belt_reaction) / params.inertia * dt;
     report.state.speed = next.clamp(-params.max_speed, params.max_speed);
-    report.state.phase = state.phase + belt_speed * dt;
+    report.state.phase = state.phase + f64::from(belt_speed * dt);
     report
 }
